ex 7
